+++ Examples/AudioTest/AudioTestGame.cs
@@ -23,8 +23,8 @@
             audioPlayer.AudioClip.AssetId = lasterId;
             audioPlayer.AddChild(new VolumeController());
 
-            scene.SaveToFile(@"TestGame - CurrentLevel.json", new Serializer());
-            this.CurrentScene = new Serializer().Deserialize<Scene>(@"TestGame - CurrentLevel.json");
+            scene.SaveToFile(@"TestGame - CurrentLevel.json");
+            this.CurrentScene = Serializer.Instance.Deserialize<Scene>(@"TestGame - CurrentLevel.json");
             this._isLoaded = true;
         }
     }