+++ Framework/Serialization/SaveDataManager.cs
@@ -7,23 +7,13 @@
     /// Managers the loading and saving of save data.
     /// </summary>
     public sealed class SaveDataManager {
-        private readonly Serializer _serializer;
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="SaveDataManager"/> class.
-        /// </summary>
-        /// <param name="serializer">The serializer.</param>
-        /// <exception cref="ArgumentNullException">serializer</exception>
-        public SaveDataManager(Serializer serializer) {
-            this._serializer = serializer ?? throw new ArgumentNullException(nameof(serializer));
-        }
 
         /// <summary>
         /// Deletes the specified save data.
         /// </summary>
         /// <param name="fileName">Name of the file.</param>
         public void Delete(string fileName) {
-            var filePath = this.GetFilePath(fileName);
+            var filePath = SaveDataManager.GetFilePath(fileName);
             if (File.Exists(filePath)) {
                 File.Delete(filePath);
             }
@@ -36,12 +26,12 @@
         /// <param name="fileName">Name of the file.</param>
         /// <returns>The save data.</returns>
         public T Load<T>(string fileName) where T : VersionedData {
-            var filePath = this.GetFilePath(fileName);
+            var filePath = SaveDataManager.GetFilePath(fileName);
             if (!File.Exists(filePath)) {
                 throw new FileNotFoundException(filePath);
             }
 
-            return this._serializer.Deserialize<T>(this.GetFilePath(fileName));
+            return Serializer.Instance.Deserialize<T>(SaveDataManager.GetFilePath(fileName));
         }
 
         /// <summary>
@@ -51,10 +41,10 @@
         /// <param name="fileName">Name of the file.</param>
         /// <param name="saveData">The save data.</param>
         public void Save<T>(string fileName, T saveData) where T : VersionedData {
-            this._serializer.Serialize(saveData, this.GetFilePath(fileName));
+            Serializer.Instance.Serialize(saveData, SaveDataManager.GetFilePath(fileName));
         }
 
-        private string GetFilePath(string fileName) {
+        private static string GetFilePath(string fileName) {
             // TODO: platforms other than windows
             var appDataDirectory = Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
             return Path.Combine(appDataDirectory, GameSettings.Instance.ProjectName.ToSafeString(), fileName);